+++ src/zeus_risk/exceptions/__init__.py
@@ -3,5 +3,11 @@
 from zeus_risk.exceptions.analytics import AnalyticsError
 from zeus_risk.exceptions.market_data import MarketDataError
 from zeus_risk.exceptions.portfolio import PortfolioImportError
+from zeus_risk.exceptions.risk import RiskCalculationError
 
-__all__ = ["AnalyticsError", "MarketDataError", "PortfolioImportError"]
+__all__ = [
+    "AnalyticsError",
+    "MarketDataError",
+    "PortfolioImportError",
+    "RiskCalculationError",
+]